+++ scheduler/cloudformation/clients.go
@@ -3,6 +3,7 @@ package cloudformation
 import (
 	"time"
 
+	awswaiter "github.com/aws/aws-sdk-go/private/waiter"
 	"github.com/aws/aws-sdk-go/service/ecs"
 	"github.com/pmylund/go-cache"
 	"github.com/remind101/empire/pkg/arn"
@@ -30,7 +31,7 @@ type cachingECSClient struct {
 }
 
 // ecsWithCaching wraps an ecs.ECS client with caching.
-func ecsWithCaching(ecs *ecs.ECS) *cachingECSClient {
+func ecsWithCaching(ecs *ECS) *cachingECSClient {
 	return &cachingECSClient{
 		ecsClient:       ecs,
 		taskDefinitions: cache.New(defaultExpiration, defaultPurge),
@@ -58,4 +59,46 @@ func (c *cachingECSClient) DescribeTaskDefinition(input *ecs.DescribeTaskDefinit
 	c.taskDefinitions.Set(*resp.TaskDefinition.TaskDefinitionArn, resp.TaskDefinition, 0)
 
 	return resp, err
+}
+
+// ECS augments the ecs.ECS client with extra waiters.
+type ECS struct {
+	*ecs.ECS
+}
+
+// WaitUntilTasksNotPending waits until all the given tasks are either RUNNING
+// or STOPPED.
+func (c *ECS) WaitUntilTasksNotPending(input *ecs.DescribeTasksInput) error {
+	waiterCfg := awswaiter.Config{
+		Operation:   "DescribeTasks",
+		Delay:       6,
+		MaxAttempts: 100,
+		Acceptors: []awswaiter.WaitAcceptor{
+			{
+				State:    "failure",
+				Matcher:  "pathAny",
+				Argument: "failures[].reason",
+				Expected: "MISSING",
+			},
+			{
+				State:    "success",
+				Matcher:  "pathAll",
+				Argument: "tasks[].lastStatus",
+				Expected: "RUNNING",
+			},
+			{
+				State:    "success",
+				Matcher:  "pathAll",
+				Argument: "tasks[].lastStatus",
+				Expected: "STOPPED",
+			},
+		},
+	}
+
+	w := awswaiter.Waiter{
+		Client: c.ECS,
+		Input:  input,
+		Config: waiterCfg,
+	}
+	return w.Wait()
 }